+++ jest.config.mjs
@@ -4,5 +4,9 @@ export default {
     '\\.ya?ml$': '<rootDir>/test/yaml-transform.mjs',
     '\\.js$': 'babel-jest',
   },
+  moduleNameMapper: {
+    // Prevent jest from picking up the ESM version.
+    '^uuid$': '<rootDir>/node_modules/uuid/dist/index.js',
+  },
   setupFilesAfterEnv: ['@testing-library/jest-dom'],
 };