+++ src/components/MediaList/AddToPlaylistAction.js
@@ -1,3 +1,4 @@
+import omit from 'just-omit';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
@@ -10,19 +11,29 @@ const {
   useCallback,
 } = React;
 
-function AddToPlaylistAction({ media }) {
+/**
+ * @param {object} props
+ * @param {object} props.media
+ * @param {boolean} [props.withCustomMeta] - Use the artist/title that is stored on this item.
+ *   Set to false when dealing with eg. search results, to make the server look up a default
+ *   artist/title for this media.
+ */
+function AddToPlaylistAction({ media, withCustomMeta = true }) {
   const { selection } = useMediaListContext();
 
   const dispatch = useDispatch();
   const handleClick = useCallback((event) => {
-    const selectedItems = selection.isSelected(media) ? selection.get() : [media];
+    let selectedItems = selection.isSelected(media) ? selection.get() : [media];
+    if (!withCustomMeta) {
+      selectedItems = selectedItems.map((item) => omit(item, ['artist', 'title']));
+    }
     const rect = event.target.getBoundingClientRect();
 
     dispatch(addMediaMenu(selectedItems, {
       x: rect.left,
       y: rect.top,
     }));
-  }, [dispatch, selection, media]);
+  }, [dispatch, selection, media, withCustomMeta]);
 
   return (
     <MediaAction onClick={handleClick}>
@@ -33,6 +44,7 @@ function AddToPlaylistAction({ media }) {
 
 AddToPlaylistAction.propTypes = {
   media: PropTypes.object.isRequired,
+  withCustomMeta: PropTypes.bool,
 };
 
 export default AddToPlaylistAction;