+++ src/constants/DDItemTypes.js
@@ -1,3 +1,4 @@
 export const MEDIA = 'dd/MEDIA';
+export const SEARCH_RESULT = 'dd/SEARCH_RESULT';
 export const HISTORY_ENTRY = 'dd/HISTORY_ENTRY';
 export const PLAYLIST = 'dd/PLAYLIST';