+++ src/sources/youtube/Player.js
@@ -1,7 +1,9 @@
 import cx from 'clsx';
 import React from 'react';
 import PropTypes from 'prop-types';
-import YouTubePlayerEmbed from './PlayerEmbed';
+import YouTube from '@u-wave/react-youtube';
+
+const { useCallback } = React;
 
 const YouTubePlayer = ({
   active,
@@ -14,17 +16,33 @@ const YouTubePlayer = ({
   onPlay,
 }) => {
   const modeClass = `src-youtube-Player--${mode}`;
+  const controllable = mode === 'preview';
+
+  const handlePause = useCallback((event) => {
+    if (active && !controllable) {
+      event.target.playVideo();
+    }
+  }, [active, controllable]);
 
   return (
     <div className={cx('src-youtube-Player', modeClass, className)} hidden={!active}>
       {enabled && (
-        <YouTubePlayerEmbed
-          media={media}
-          active={active}
-          seek={Math.round(seek)}
-          volume={volume}
-          controllable={mode === 'preview'}
-          onPlay={onPlay}
+        <YouTube
+          video={active ? media.sourceID : null}
+          width="100%"
+          height="100%"
+          autoplay
+          modestBranding
+          disableKeyboard
+          volume={volume / 100}
+          playbackRate={1}
+          controls={controllable}
+          showRelatedVideos={false}
+          annotations={false}
+          startSeconds={Math.round(seek + (media.start ?? 0))}
+          endSeconds={media.end ?? media.duration}
+          onPause={handlePause}
+          onPlaying={onPlay}
         />
       )}
     </div>