+++ test/yaml-transform.mjs
@@ -2,7 +2,7 @@ import yaml from 'js-yaml';
 
 function process(src) {
   const json = yaml.load(src);
-  return `module.exports = ${JSON.stringify(json)};`;
+  return { code: `module.exports = ${JSON.stringify(json)};` };
 }
 
 export default { process };